requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
cmake==3.30.3
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.0
idna==3.10
Jinja2==3.1.4
lit==18.1.8
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.1
pillow==10.4.0
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.20.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.5
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
